Add Event: Then: I must see the message {string}
This requires us to submit the form, so we can assert if a message is
displayed.  We also assert that a new event is added, which we use an
existing Then step for :)

The implementation here is more complicated than it should, as I realize
we need some way of refreshing our list of events, which is in a
different component other than AddEvent.  We'll sort this out in the
next set of commits.