npm run deploy
It seems that React refers to the files in distDir, probably for
incremental build.  This is a problem, as we clear api/public as we can
deploy either Vue or React into it.

Solution is to manually copy the output to api/public, which is the same
approach we do for Vite.

Also made sure to clear the contents of api/public before copying into
it.  Interestingly, this caused a problem for Vite, as favicon.ico was
now missing!  Solved that by adding it to app/public :)